Renames checkForKeywordInUrl to countKeywordInUrl
